+++ test/integration/shop/b113_s1_profile_header_test.rb
@@ -40,7 +40,6 @@ class Shop::B113S1ProfileHeaderTest < ActionDispatch::IntegrationTest
     refute_includes bottom, ">Профиль<"
     refute_includes bottom, "push('/profile')"
     assert_includes bottom, "Каталог"
-    assert_includes bottom, "Корзина"
     assert_includes bottom, "Избранное"
   end